feat: support 'tap' gesture for Gesture plugin
